Resolved linter PR review comments
